Replace parity bits with CRC

+++ LaserTag.X/IRReceiver.h
@@ -13,6 +13,6 @@ void receiverEventHandler(void);
 // transmission was received since the last call to tryGetTransmissionData.
 // Returns false otherwise. The contents of data_out are undefined when this
 // function returns false.
-bool tryGetTransmissionData(uint16_t* data_out);
+bool tryGetTransmissionData(uint8_t* data_out);
 
 #endif	/* IRRECEIVER_H */
