+++ LaserTag.X/IRTransmitter.h
@@ -11,6 +11,6 @@ void transmitterEventHandler(void);
 
 // Asynchronously transmit the given bits over IR. If a transmission is in
 // progress, returns false and does nothing. Otherwise returns true
-bool transmitAsync(uint16_t data);
+bool transmitAsync(uint8_t data);
 
 #endif	/* IRTRANSMITTER_H */
